agricultor/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pika
cryptography